# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_post",
        "type": "object"
      },
...
      "CoverageOut": {
        "description": "What the reading did with one part of the file: read, accepted, excluded, partial, omitted, absent or unknown.",
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "part": {
            "title": "Part",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "part",
          "status"
        ],
        "title": "CoverageOut",
        "type": "object"
      },
      "CoverageReport": {
        "description": "A reading's coverage with its provenance. persisted: recorded by the reading at upload. reconstructed: computed later\nfrom the stored bytes under the same reader version; the same facts, not captured at the time.",
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/CoverageOut"
            },
            "title": "Parts",
            "type": "array"
          },
          "readerVersion": {
            "title": "Readerversion",
            "type": "string"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "source": {
            "enum": [
              "persisted",
              "reconstructed"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "schemaVersion",
          "readerVersion",
          "source",
          "parts"
        ],
        "title": "CoverageReport",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoverageReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "default": "",
            "title": "Createdat",
            "type": "string"
          },
          "hiddenRuns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hiddenruns"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "parseMs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsems"
          },
          "parserVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parserversion"
          },
          "reused": {
            "default": false,
            "title": "Reused",
            "type": "boolean"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "trackedChanges": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackedchanges"
          }
        },
        "required": [
          "id",
          "name",
          "pages",
          "sha256",
          "sections"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "DocumentSummary": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "findings": {
            "default": 0,
            "title": "Findings",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastRunAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastrunat"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "reviewedFindings": {
            "default": 0,
            "title": "Reviewedfindings",
            "type": "integer"
          },
          "sections": {
            "title": "Sections",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "pages",
          "sections",
          "createdAt"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
...
      "FindingRecord": {
        "properties": {
          "citations": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "conclusion": {
            "title": "Conclusion",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "documentId": {
            "title": "Documentid",
            "type": "string"
          },
          "documentName": {
            "title": "Documentname",
            "type": "string"
          },
          "evidenceKind": {
            "default": "passage",
            "enum": [
              "passage",
              "coverage"
            ],
            "title": "Evidencekind",
            "type": "string"
          },
          "hasGuidance": {
            "default": false,
            "title": "Hasguidance",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "shared": {
            "default": false,
            "title": "Shared",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "pass",
              "needs_review",
              "missing",
              "unresolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "statusSource": {
            "anyOf": [
              {
                "enum": [
                  "computed_days",
                  "confirmed_days",
                  "model_hint",
                  "no_evidence",
                  "reference_check",
                  "position_check",
                  "ambiguous_fact"
                ],
                "type": "string"
              },
              {
                "const": "",
                "type": "string"
              }
            ],
            "default": "model_hint",
            "title": "Statussource"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "verifiedSpans": {
            "title": "Verifiedspans",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "runId",
          "documentId",
          "documentName",
          "question",
          "topic",
          "status",
          "conclusion",
          "verifiedSpans",
          "createdAt"
        ],
        "title": "FindingRecord",
        "type": "object"
      },
      "FindingReviewOut": {
        "properties": {
          "findingId": {
            "title": "Findingid",
            "type": "string"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "findingId"
        ],
        "title": "FindingReviewOut",
        "type": "object"
      },
...
      "GuidanceIn": {
        "properties": {
          "text": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "GuidanceIn",
        "type": "object"
      },
      "GuidanceOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sha256",
          "text"
        ],
        "title": "GuidanceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "confirmed",
              "dismissed",
              "cleared"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict",
          "reviewer"
        ],
        "title": "ReviewIn",
        "type": "object"
      },
      "ReviewOut": {
        "description": "A person's current decision on a finding; absent when unreviewed or cleared.",
        "properties": {
          "accessSubject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accesssubject"
          },
          "at": {
            "title": "At",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "confirmed",
              "dismissed"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict",
          "reviewer",
          "at"
        ],
        "title": "ReviewOut",
        "type": "object"
      },
...
      "SectionOut": {
        "properties": {
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "number",
          "heading",
          "text"
        ],
        "title": "SectionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Contract Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "description": "This workspace's documents and the curated ones; the counts beside each are of runs it may read.",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "title": "Response List Documents Api Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "201 with a document new to this workspace, or 200 with the one it already holds that has exactly these bytes.",
        "operationId": "upload_document_api_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/findings": {
      "get": {
        "description": "Findings of the runs this workspace may read: its own and the curated ones.",
        "operationId": "list_findings_api_findings_get",
        "parameters": [
          {
            "in": "query",
            "name": "documentId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Documentid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FindingRecord"
                  },
                  "title": "Response List Findings Api Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Findings",
        "tags": [
          "findings"
        ]
      }
    },
    "/api/findings/{finding_id}/review": {
      "post": {
        "description": "Record a person's decision. 201 when it changed the state, 200 when it repeated it; `cleared` undoes. With the gate\non, the invite's subject is recorded beside the typed name; the name is what the browser sent and is not an identity.",
        "operationId": "review_api_findings__finding_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingReviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review",
        "tags": [
          "findings"
        ]
      }
    },
    "/api/guidance": {
      "post": {
        "description": "201 with a new guidance record in this workspace, or 200 with the one it may already use that holds these words.",
        "operationId": "create_guidance_api_guidance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidanceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Guidance",
        "tags": [
          "guidance"
        ]
      }
    },
    "/api/guidance/{guidance_id}": {
      "get": {
        "operationId": "get_guidance_api_guidance__guidance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "guidance_id",
            "required": true,
            "schema": {
              "title": "Guidance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Guidance",
        "tags": [
          "guidance"
        ]
      }
    },
...
  }
}